requirements.txt
numpy
matplotlib
streamlit
streamlit-shortcuts
st-gsheets-connection
validators==0.22.0
google-cloud-firestore
